Move OMC and NotifyNL Test Endpoints to own controller And refactor
