+++ OMC/Core/Domain/EventsHandler/Attributes/Validation/StandardizeApiResponsesAttribute.cs
@@ -32,7 +32,9 @@ namespace EventsHandler.Attributes.Validation
             // NOTE: Concept similar to strategy design pattern => decide how and which API Controllers are responding to the end-user
             s_mappedControllersToResponders.TryAdd(typeof(EventsController), typeof(OmcResponder));
             s_mappedControllersToResponders.TryAdd(typeof(NotifyController), typeof(NotifyResponder));
-            s_mappedControllersToResponders.TryAdd(typeof(TestController), typeof(NotifyResponder));
+            s_mappedControllersToResponders.TryAdd(typeof(TestNotifyNLController), typeof(NotifyResponder));
+            s_mappedControllersToResponders.TryAdd(typeof(TestZHVController), typeof(NotifyResponder));
+            s_mappedControllersToResponders.TryAdd(typeof(TestOMCController), typeof(NotifyResponder));
         }
 
         /// <summary>
